requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic>=2.0.0
pydantic_core
jinja2
Babel
python-multipart
itsdangerous

# Segurança
passlib[bcrypt]
python-jose[cryptography]
python-dotenv>=1.0.0
slowapi>=0.1.9
fastapi-csrf-protect>=0.3.4
email-validator>=2.0.0

# Processamento de imagens e uploads
Pillow>=10.0.0

# Dependências de teste
pytest
pytest-asyncio
pytest-cov